update template and test-utils
