+++ sdk/template/template/package.json
@@ -79,7 +79,7 @@
   "dependencies": {
     "@azure/core-auth": "^1.3.0",
     "@azure/core-http": "^2.0.0",
-    "@azure/core-tracing": "1.0.0-preview.13",
+    "@azure/core-tracing": "1.0.0-preview.14",
     "@azure/logger": "^1.0.0",
     "tslib": "^2.2.0"
   },
@@ -87,6 +87,8 @@
     "@azure/dev-tool": "^1.0.0",
     "@azure/eslint-plugin-azure-sdk": "^3.0.0",
     "@azure/identity": "^2.0.1",
+    "@azure/test-utils": "^1.0.0",
+    "@azure-tools/test-credential": "~1.0.0",
     "@azure-tools/test-recorder": "^2.0.0",
     "@microsoft/api-extractor": "^7.18.11",
     "@types/chai": "^4.1.6",
@@ -121,8 +123,7 @@
     "rollup": "^1.16.3",
     "source-map-support": "^0.5.9",
     "typescript": "~4.2.0",
-    "util": "^0.12.1",
-    "@azure-tools/test-credential": "~1.0.0"
+    "util": "^0.12.1"
   },
   "//sampleConfiguration": {
     "skipFolder": false,
