+++ sdk/template/template/recordings/browsers/aad_configurationclient_functional_tests_getconfigurationsetting/recording_predetermined_setting_has_expected_value.json
@@ -5,12 +5,20 @@
       "RequestMethod": "GET",
       "RequestHeaders": {
         "Accept": "application/vnd.microsoft.appconfig.kv\u002Bjson, application/json, application/problem\u002Bjson",
-        "Accept-Encoding": "gzip,deflate",
+        "Accept-Encoding": "gzip, deflate, br",
+        "Accept-Language": "en-US",
         "Authorization": "Sanitized",
         "Connection": "keep-alive",
-        "Cookie": "",
-        "User-Agent": "azsdk-js-template-template/1.0.11-beta.1 core-http/2.2.3 Node/v16.13.0 OS/(x64-Linux-5.10.60.1-microsoft-standard-WSL2)",
-        "x-ms-client-request-id": "f6f68ce8-5b02-461d-bdf1-2af09b79624a"
+        "Referer": "http://localhost:9876/",
+        "sec-ch-ua": "",
+        "sec-ch-ua-mobile": "?0",
+        "sec-ch-ua-platform": "",
+        "Sec-Fetch-Dest": "empty",
+        "Sec-Fetch-Mode": "cors",
+        "Sec-Fetch-Site": "same-site",
+        "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/93.0.4577.0 Safari/537.36",
+        "x-ms-client-request-id": "25fc4075-1b16-4329-8515-80466453103d",
+        "x-ms-useragent": "azsdk-js-template-template/1.0.11-beta.1 core-http/2.2.4 OS/Linuxx86_64"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -20,25 +28,25 @@
         "Access-Control-Expose-Headers": "DNT, X-CustomHeader, Keep-Alive, User-Agent, X-Requested-With, If-Modified-Since, Cache-Control, Content-Type, Authorization, x-ms-client-request-id, x-ms-useragent, x-ms-content-sha256, x-ms-date, host, Accept, Accept-Datetime, Date, If-Match, If-None-Match, Sync-Token, x-ms-return-client-request-id, ETag, Last-Modified, Link, Memento-Datetime, retry-after-ms, x-ms-request-id, x-ms-client-session-id, x-ms-effective-locale, WWW-Authenticate, traceparent, tracestate",
         "Connection": "keep-alive",
         "Content-Type": "application/vnd.microsoft.appconfig.kv\u002Bjson; charset=utf-8",
-        "Date": "Thu, 09 Dec 2021 01:06:27 GMT",
-        "ETag": "\u0022jRXE6grlhA98vbDBE98P9JDaMPm\u0022",
-        "Last-Modified": "Thu, 09 Dec 2021 01:05:58 GMT",
+        "Date": "Fri, 28 Jan 2022 18:12:31 GMT",
+        "ETag": "\u0022jHnSsNSkboEWJFD7nnfzHgnZjRA\u0022",
+        "Last-Modified": "Wed, 01 Dec 2021 02:23:41 GMT",
         "Server": "openresty/1.17.8.2",
         "Strict-Transport-Security": "max-age=15724800; includeSubDomains",
-        "Sync-Token": "zAJw6V16=NToxOSM4NDcxODQ4;sn=8471848",
+        "Sync-Token": "zAJw6V16=NDoxOSM1ODg2MzAy;sn=5886302",
         "Transfer-Encoding": "chunked",
-        "x-ms-correlation-request-id": "5625011c-719c-4adc-ab86-a0b59fda97dd",
-        "x-ms-request-id": "5625011c-719c-4adc-ab86-a0b59fda97dd"
+        "x-ms-correlation-request-id": "b93931e9-2d38-407f-b20c-8516018db052",
+        "x-ms-request-id": "b93931e9-2d38-407f-b20c-8516018db052"
       },
       "ResponseBody": {
-        "etag": "jRXE6grlhA98vbDBE98P9JDaMPm",
+        "etag": "jHnSsNSkboEWJFD7nnfzHgnZjRA",
         "key": "test-key",
         "label": null,
-        "content_type": "",
+        "content_type": null,
         "value": "test-value",
         "tags": {},
         "locked": false,
-        "last_modified": "2021-12-09T01:05:58\u002B00:00"
+        "last_modified": "2021-12-01T02:23:41\u002B00:00"
       }
     }
   ],
