+++ sdk/template/template/recordings/browsers/aad_configurationclient_functional_tests_getconfigurationsetting/recording_supports_tracing.json
@@ -1,0 +1,4 @@
+{
+  "Entries": [],
+  "Variables": {}
+}
