+++ sdk/template/template/recordings/node/aad_configurationclient_functional_tests_getconfigurationsetting/recording_predetermined_setting_has_expected_value.json
@@ -1,0 +1,46 @@
+{
+  "Entries": [
+    {
+      "RequestUri": "https://myappconfig.azconfig.io/kv/test-key?api-version=1.0",
+      "RequestMethod": "GET",
+      "RequestHeaders": {
+        "Accept": "application/vnd.microsoft.appconfig.kv\u002Bjson, application/json, application/problem\u002Bjson",
+        "Accept-Encoding": "gzip,deflate",
+        "Authorization": "Sanitized",
+        "Connection": "keep-alive",
+        "Cookie": "",
+        "User-Agent": "azsdk-js-template-template/1.0.11-beta.1 core-http/2.2.4 Node/v16.13.1 OS/(x64-Linux-5.10.60.1-microsoft-standard-WSL2)",
+        "x-ms-client-request-id": "9794224c-ab46-44c5-a2d6-c553f15d9e54"
+      },
+      "RequestBody": null,
+      "StatusCode": 200,
+      "ResponseHeaders": {
+        "Access-Control-Allow-Credentials": "true",
+        "Access-Control-Allow-Origin": "*",
+        "Access-Control-Expose-Headers": "DNT, X-CustomHeader, Keep-Alive, User-Agent, X-Requested-With, If-Modified-Since, Cache-Control, Content-Type, Authorization, x-ms-client-request-id, x-ms-useragent, x-ms-content-sha256, x-ms-date, host, Accept, Accept-Datetime, Date, If-Match, If-None-Match, Sync-Token, x-ms-return-client-request-id, ETag, Last-Modified, Link, Memento-Datetime, retry-after-ms, x-ms-request-id, x-ms-client-session-id, x-ms-effective-locale, WWW-Authenticate, traceparent, tracestate",
+        "Connection": "keep-alive",
+        "Content-Type": "application/vnd.microsoft.appconfig.kv\u002Bjson; charset=utf-8",
+        "Date": "Fri, 28 Jan 2022 18:12:21 GMT",
+        "ETag": "\u0022jHnSsNSkboEWJFD7nnfzHgnZjRA\u0022",
+        "Last-Modified": "Wed, 01 Dec 2021 02:23:41 GMT",
+        "Server": "openresty/1.17.8.2",
+        "Strict-Transport-Security": "max-age=15724800; includeSubDomains",
+        "Sync-Token": "zAJw6V16=NDoxOSM1ODg2Mjk4;sn=5886298",
+        "Transfer-Encoding": "chunked",
+        "x-ms-correlation-request-id": "9cfed928-f387-4795-93e1-0dfa7ae6d3b4",
+        "x-ms-request-id": "9cfed928-f387-4795-93e1-0dfa7ae6d3b4"
+      },
+      "ResponseBody": {
+        "etag": "jHnSsNSkboEWJFD7nnfzHgnZjRA",
+        "key": "test-key",
+        "label": null,
+        "content_type": null,
+        "value": "test-value",
+        "tags": {},
+        "locked": false,
+        "last_modified": "2021-12-01T02:23:41\u002B00:00"
+      }
+    }
+  ],
+  "Variables": {}
+}
