+++ sdk/template/template/recordings/node/aad_configurationclient_functional_tests_getconfigurationsetting/recording_supports_tracing.json
@@ -5,20 +5,12 @@
       "RequestMethod": "GET",
       "RequestHeaders": {
         "Accept": "application/vnd.microsoft.appconfig.kv\u002Bjson, application/json, application/problem\u002Bjson",
-        "Accept-Encoding": "gzip, deflate, br",
-        "Accept-Language": "en-US",
+        "Accept-Encoding": "gzip,deflate",
         "Authorization": "Sanitized",
         "Connection": "keep-alive",
-        "Referer": "http://localhost:9876/",
-        "sec-ch-ua": "",
-        "sec-ch-ua-mobile": "?0",
-        "sec-ch-ua-platform": "",
-        "Sec-Fetch-Dest": "empty",
-        "Sec-Fetch-Mode": "cors",
-        "Sec-Fetch-Site": "same-site",
-        "User-Agent": "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) HeadlessChrome/93.0.4577.0 Safari/537.36",
-        "x-ms-client-request-id": "0a240cc2-fffd-45da-9a4c-fe8cebdff715",
-        "x-ms-useragent": "azsdk-js-template-template/1.0.11-beta.1 core-http/2.2.3 OS/Linuxx86_64"
+        "Cookie": "",
+        "User-Agent": "azsdk-js-template-template/1.0.11-beta.1 core-http/2.2.4 Node/v16.13.1 OS/(x64-Linux-5.10.60.1-microsoft-standard-WSL2)",
+        "x-ms-client-request-id": "2d164b86-014c-4013-9c8c-56aa7883e56e"
       },
       "RequestBody": null,
       "StatusCode": 200,
@@ -28,15 +20,15 @@
         "Access-Control-Expose-Headers": "DNT, X-CustomHeader, Keep-Alive, User-Agent, X-Requested-With, If-Modified-Since, Cache-Control, Content-Type, Authorization, x-ms-client-request-id, x-ms-useragent, x-ms-content-sha256, x-ms-date, host, Accept, Accept-Datetime, Date, If-Match, If-None-Match, Sync-Token, x-ms-return-client-request-id, ETag, Last-Modified, Link, Memento-Datetime, retry-after-ms, x-ms-request-id, x-ms-client-session-id, x-ms-effective-locale, WWW-Authenticate, traceparent, tracestate",
         "Connection": "keep-alive",
         "Content-Type": "application/vnd.microsoft.appconfig.kv\u002Bjson; charset=utf-8",
-        "Date": "Fri, 10 Dec 2021 11:09:32 GMT",
+        "Date": "Fri, 28 Jan 2022 18:12:22 GMT",
         "ETag": "\u0022jHnSsNSkboEWJFD7nnfzHgnZjRA\u0022",
         "Last-Modified": "Wed, 01 Dec 2021 02:23:41 GMT",
         "Server": "openresty/1.17.8.2",
         "Strict-Transport-Security": "max-age=15724800; includeSubDomains",
-        "Sync-Token": "zAJw6V16=NDoxOSM1MjM0MTM4;sn=5234138",
+        "Sync-Token": "zAJw6V16=NDoxOSM1ODg2Mjk4;sn=5886298",
         "Transfer-Encoding": "chunked",
-        "x-ms-correlation-request-id": "cb5d6844-d7dd-48e7-866f-dcb2331ed70f",
-        "x-ms-request-id": "cb5d6844-d7dd-48e7-866f-dcb2331ed70f"
+        "x-ms-correlation-request-id": "9e6dc18d-5d7f-4b20-a31d-bff95437a0ea",
+        "x-ms-request-id": "9e6dc18d-5d7f-4b20-a31d-bff95437a0ea"
       },
       "ResponseBody": {
         "etag": "jHnSsNSkboEWJFD7nnfzHgnZjRA",
