+++ sdk/template/template/src/configurationClient.ts
@@ -10,9 +10,8 @@ import {
   bearerTokenAuthenticationPolicy,
   createPipelineFromOptions,
 } from "@azure/core-http";
+import { TracingClient, createTracingClient } from "@azure/core-tracing";
 import { SDK_VERSION } from "./constants";
-import { SpanStatusCode } from "@azure/core-tracing";
-import { createSpan } from "./tracing";
 import { logger } from "./logger";
 import { quoteETag } from "./util";
 
@@ -46,6 +45,7 @@ export interface ConfigurationClientOptions extends PipelineOptions {
  */
 export class ConfigurationClient {
   private client: GeneratedClient;
+  private tracingClient: TracingClient;
 
   /**
    * Creates an instance of a ConfigurationClient.
@@ -109,6 +109,14 @@ export class ConfigurationClient {
     const pipeline = createPipelineFromOptions(internalPipelineOptions, authPolicy);
 
     this.client = new GeneratedClient(endpointUrl, pipeline);
+    this.tracingClient = createTracingClient({
+      // The name of the resource provider requests are made against, as described in
+      // https://github.com/Azure/azure-sdk/blob/main/docs/tracing/distributed-tracing-conventions.yml#L11-L15
+      namespace: "Microsoft.Learn",
+      // The package name and version
+      packageName: "@azure/template",
+      packageVersion: SDK_VERSION,
+    });
   }
 
   /**
@@ -141,41 +149,29 @@ export class ConfigurationClient {
     let key: string;
     let ifNoneMatch: string | undefined;
 
-    if (typeof keyOrSetting === "string") {
-      key = keyOrSetting;
-      if (options.onlyIfChanged) {
-        throw new RangeError(
-          "You must pass a ConfigurationSetting instead of a key to perform a conditional fetch."
-        );
-      }
-    } else {
-      key = keyOrSetting.key;
-      const etag = keyOrSetting.etag;
-      if (options.onlyIfChanged) {
-        ifNoneMatch = quoteETag(etag);
-      }
-    }
+    return this.tracingClient.withSpan(
+      // Span names should take the form "<className>.<methodName>".
+      "ConfigurationClient.getConfigurationSetting",
+      options,
+      (updatedOptions) => {
+        if (typeof keyOrSetting === "string") {
+          key = keyOrSetting;
+          if (options.onlyIfChanged) {
+            throw new RangeError(
+              "You must pass a ConfigurationSetting instead of a key to perform a conditional fetch."
+            );
+          }
+        } else {
+          key = keyOrSetting.key;
+          const etag = keyOrSetting.etag;
+          if (options.onlyIfChanged) {
+            ifNoneMatch = quoteETag(etag);
+          }
+        }
 
-    const { span, updatedOptions } = createSpan(
-      // Here you set the name of the span, usually clientName-operationName
-      "ConfigurationClient-getConfigurationSetting",
-      options
+        // You must pass updatedOptions to any calls you make within the callback.
+        return this.client.getKeyValue(key, { ...updatedOptions, ifNoneMatch });
+      }
     );
-
-    try {
-      const result = await this.client.getKeyValue(key, {
-        ...updatedOptions,
-        ifNoneMatch,
-      });
-      return result;
-    } catch (e) {
-      // There are different standard codes available for different errors:
-      // https://github.com/open-telemetry/opentelemetry-specification/blob/master/specification/trace/api.md#status
-      span.setStatus({ code: SpanStatusCode.ERROR, message: e.message });
-
-      throw e;
-    } finally {
-      span.end();
-    }
   }
 }
